chatboxes: compact pill bar on Publish tab
Mirrors the evals suite header pattern: a Globe pill that opens the
shared AttachmentEditor modal for the chatbox's server pick, plus a
host pill that routes to Connect for identity edits. Replaces the
previous "Servers / Edit servers" card.
